Fix crash when trying to report crash
Sorry about that. In the original code I PRed, I accidentally tried to resolve the Structure with the dynamic registry which fails because ConfiguredStructures are in the dynamic registry. To resolve the structure properly, the structure registry should be used instead. please test this pr to make sure I got the right mappings as I am doing this edit through github's website instead of my IDE to try and get this to your attention quickly before I get to busy with irl stuff. 

Again, apologies for accidentally PRing code that crashed before which hid the actual crash. The flawed code was also in my mod so I had to fix my mod as well. If this works in your IDE and compiles and actually prints properly without crashing, be sure to port the fix to 1.17 as well and to both forge and fabric